Qualification emC for Simulink S-Func Inspc etc.
* emC/Base/+genRefl.jz.sh war empty meanwhile, restaurating from 2020-12-04
* #inlude <emC/Base/Object_emC> was not necessary in the past, was contained in applstdef_emC.h
* ObjectSimple_emC.h: size with -1 was used as "do not set", regard.
* emC/Base/genRefl/*.refl all newly generated, with new tool, TODO Test.
  The Test_emC yet not includes newly reflection generation (does not include the reflection tool test)
* Some more adaption in files, which are not part compilation in of Test_emC
* applstdef_emC.h and emC_inclComplSpec/cc_SmlkSfunc_64bit/compl_adaption.h compiled toward running, not toward all ok

+++ emC/J1c/StringPartJc.c
@@ -2447,7 +2447,7 @@ bool startsWith_StringPartJc(StringPartJc_s* thiz, CharSeqJc sCmp, ThCxt* _thCxt
 
 
 /**Gets the current position, useable for rewind. This method is overwritten*/
-int64 getCurrentPosition_StringPartJc(StringPartJc_s* thiz, ThCxt* _thCxt)
+int32 getCurrentPosition_StringPartJc(StringPartJc_s* thiz, ThCxt* _thCxt)
 { 
   STACKTRC_TENTRY("getCurrentPosition_StringPartJc");
   
@@ -2462,7 +2462,7 @@ int64 getCurrentPosition_StringPartJc(StringPartJc_s* thiz, ThCxt* _thCxt)
 
 
 /**Sets the current position at a fix position inside the maxPart.*/
-void setCurrentPosition_StringPartJc(StringPartJc_s* thiz, int64 pos, ThCxt* _thCxt)
+void setCurrentPosition_StringPartJc(StringPartJc_s* thiz, int32 pos, ThCxt* _thCxt)
 { 
   STACKTRC_TENTRY("setCurrentPosition_StringPartJc");
